+++ packages/riverpod_lint_flutter_test/test/goldens/lints/scoped_providers_should_specify_dependencies.dart
@@ -8,12 +8,16 @@ part 'scoped_providers_should_specify_dependencies.g.dart';
 int scoped(ScopedRef ref) => 0;
 
 @riverpod
+external int unimplementedScoped();
+
+@riverpod
 int root(RootRef ref) => 0;
 
 void main() {
   final rootContainer = ProviderContainer(
     overrides: [
       scopedProvider.overrideWith((ref) => 0),
+      unimplementedScopedProvider.overrideWith((ref) => 0),
       rootProvider.overrideWith((ref) => 0),
     ],
   );
@@ -23,6 +27,7 @@ void main() {
     parent: rootContainer,
     overrides: [
       scopedProvider.overrideWith((ref) => 0),
+      unimplementedScopedProvider.overrideWith((ref) => 0),
       // expect_lint: scoped_providers_should_specify_dependencies
       rootProvider.overrideWith((ref) => 0),
     ],
@@ -32,6 +37,7 @@ void main() {
     ProviderScope(
       overrides: [
         scopedProvider.overrideWith((ref) => 0),
+        unimplementedScopedProvider.overrideWith((ref) => 0),
         rootProvider.overrideWith((ref) => 0),
       ],
       child: Container(),
@@ -43,6 +49,7 @@ void main() {
       parent: rootContainer,
       overrides: [
         scopedProvider.overrideWith((ref) => 0),
+        unimplementedScopedProvider.overrideWith((ref) => 0),
         // expect_lint: scoped_providers_should_specify_dependencies
         rootProvider.overrideWith((ref) => 0),
       ],
@@ -55,6 +62,7 @@ Widget fn() {
   return ProviderScope(
     overrides: [
       scopedProvider.overrideWith((ref) => 0),
+      unimplementedScopedProvider.overrideWith((ref) => 0),
       // expect_lint: scoped_providers_should_specify_dependencies
       rootProvider.overrideWith((ref) => 0),
     ],
@@ -69,6 +77,7 @@ void showModal(BuildContext context) {
       return ProviderScope(
         overrides: [
           scopedProvider.overrideWith((ref) => 0),
+          unimplementedScopedProvider.overrideWith((ref) => 0),
           // expect_lint: scoped_providers_should_specify_dependencies
           rootProvider.overrideWith((ref) => 0),
         ],
@@ -86,6 +95,7 @@ class MyWidget extends StatelessWidget {
     return ProviderScope(
       overrides: [
         scopedProvider.overrideWith((ref) => 0),
+        unimplementedScopedProvider.overrideWith((ref) => 0),
         // expect_lint: scoped_providers_should_specify_dependencies
         rootProvider.overrideWith((ref) => 0),
       ],