+++ packages/riverpod_lint/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased fix
 
 - Fix false positive with `stateless_ref` lint on scoped providers
+- Fix false positive with `provider_dependencies` lint on scoped providers
 
 # 1.1.0